1.前端搜索完成，实现按栏目搜索，关键词搜索 2.删除没必要的文件，和数据库字段，如栏目排序，删除验证规则

+++ application/admin/controller/Category.php
@@ -45,6 +45,7 @@ class Category extends Common{
         $id = input('param.id');
        
         if (request()->isPost()) {
+            /*halt(input('post.'));*/
            $res = $this->db->edit(input('post.'));
             if ($res['valid']) {
                 $this->success($res['msg'],'index');
@@ -67,6 +68,11 @@ class Category extends Common{
     //删除
     public function delete(){
          $id = input('param.id');
+         $pid = $this->db->where('id',$id)->find();
+        // halt($pid['pid']);
+         if ($pid['pid'] == 0) {
+            return json(['valid'=>2,'msg'=>'该栏目为顶级栏目不允许删除！']);
+         }
          if(CategoryModel::destroy($id)){
             return json(['valid'=>1,'msg'=>'删除成功']);
          }else{
